Fix EFR32 format string and ESP32 missing clusters

+++ examples/all-devices-app/all-devices-common/devices/water-valve/WaterValveDevice.cpp
@@ -93,7 +93,7 @@ CHIP_ERROR WaterValveDevice::HandleCloseValve()
 
 void WaterValveDevice::HandleRemainingDurationTick(uint32_t duration)
 {
-    ChipLogProgress(AppServer, "WaterValveDevice: Remaining duration tick: %u", duration);
+    ChipLogProgress(AppServer, "WaterValveDevice: Remaining duration tick: %lu", static_cast<unsigned long>(duration));
 }
 
 } // namespace chip::app
